% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGRscreen.R
\name{eGRscreen}
\alias{eGRscreen}
\title{Uses eGR to screen potential biomarker genes}
\usage{
eGRscreen(
  tv.data,
  ref.group,
  model.var,
  group.var,
  mouse.id.var,
  time.var,
  tv.var,
  gene.expr.data,
  type = c("difference", "ratio")
)
}
\arguments{
\item{tv.data}{tv data frame in the "long" format, must contain a "Model"
column that matches the sample names of gene expression data. Must have mouse ID
variable that is unique at least within the experimental group.}

\item{ref.group}{The reference group, usually a control/vehicle group for the drug group to compare with. Required.}

\item{model.var}{The name of the mouse model column. Required.}

\item{group.var}{The name of the group column. Required.}

\item{mouse.id.var}{The name of the mouse ID column. Required.}

\item{time.var}{The name of the time column. Required.}

\item{tv.var}{The name of the TV column. Required.}

\item{gene.expr.data}{gene expression data in the "fat" format, in which every column is a gene, and every row is a mouse model.
The first column must be the sample names.Required.}
}
\value{
a list of eGR difference or ratio result for each model, and gene list ordered by p-value
}
\description{
Uses eGR to screen potential biomarker genes
}
