% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamFit.R
\name{gamFit}
\alias{gamFit}
\title{Fit tumor growth curves using generalized additive model (GAM)}
\usage{
gamFit(tv.data, ref.group, group.id.var, mouse.id.var)
}
\arguments{
\item{tv.data}{Tumor volume data in "long format", that is, tumor volumes are in separate lines, with a Day column. Required.}

\item{ref.group}{The reference group, usually a control/vehicle group for the drug group to compare with. Required.}

\item{group.id.var}{The name of the "group" column of your input data. Required.}

\item{mouse.id.var}{The name of the "mouse.id" column of your input data. Required.}
}
\value{
An S3 object of class "gamfit" which stores coefficient tables and a "gam" object of the fitted GAM model.
}
\description{
This function uses generalized additive model to test if there are significant differences in growth rates between treatment groups.
}
\details{
Generalized additive model fits tumor growth curves with smooth functions (with regularization to avoid over-fitting).
p-value is given to test whether there is significant difference between the intercepts of groups (which can be a surrogate for tumor growth rates)
}
