% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamscreen.R
\name{gamscreen}
\alias{gamscreen}
\title{Uses generalized additive mixed model to screen potential biomarker genes}
\usage{
gamscreen(
  tv.data,
  ref.group,
  model.var,
  group.var,
  mouse.id.var,
  gene.expr.data
)
}
\arguments{
\item{tv.data}{tv data frame in the "long" format, must contain a "Model"
column that matches the sample names of gene expression data. Must have mouse ID variable that is unique at least within the experimental group.}

\item{ref.group}{The reference group, usually a control/vehicle group for the drug group to compare with. Required.}

\item{model.var}{The name of the mouse model column. Do not add quotation marks.}

\item{group.var}{The name of the group column. Do not add quotation marks.}

\item{mouse.id.var}{The name of the mouse ID column. Do not add quotation marks.}

\item{gene.expr.data}{gene expression data in the "fat" format, in which every column is a gene, and every row is a mouse model.
The first column must be the sample names.}
}
\description{
Uses GAMM to screen potential biomarker genes one by one. It quotes gamm4() function from gamm4 package to fit growth curves.
}
