% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eGR.R
\name{get_eGR}
\alias{get_eGR}
\title{calculate eGR for a single tumor}
\usage{
get_eGR(df)
}
\arguments{
\item{df}{a single tumor volume data frame in long format, which contains columns  "Group","Mouse","Day", "TV"}
}
\value{
a data frame of eGR information
}
\description{
calculate eGR for a single tumor
}
