% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_eGR.R
\name{get_group_eGR}
\alias{get_group_eGR}
\title{get eGR for a single group of tumors}
\usage{
get_group_eGR(tv.data, mouse.id.var, time.var, tv.var, ci = F, nrep = 1000)
}
\arguments{
\item{tv.data}{tumor volume data in long format, containing columns "Group","Mouse","Day","TV"}

\item{mouse.id.var}{The name of the "mouse.id" column of your input data. Required.}

\item{time.var}{The name of the "day" column of your input data. Required.}

\item{tv.var}{The name of the "tumor volume" column of your input data. Required.}

\item{ci}{a logical variable to do bootstrapping for 90% confidence interval}

\item{nrep}{the number of bootstrapping for confidence interval}
}
\value{
a data frame containing eGR for this group
}
\description{
get eGR for a single group of tumors
}
