% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Model_eGR.R
\name{get_Model_eGR}
\alias{get_Model_eGR}
\title{get eGR ratio or difference within a trial}
\usage{
get_Model_eGR(
  tv.data,
  ref.group,
  group.id.var,
  mouse.id.var,
  time.var,
  tv.var,
  ci = F,
  type = c("difference", "ratio")
)
}
\arguments{
\item{tv.data}{tumor volume data in long format, containing columns "Group","Mouse","Day","TV"}

\item{ref.group}{the reference group in eGR ratio or difference}

\item{group.id.var}{The name of the "group" column of your input data. Required.}

\item{mouse.id.var}{The name of the "mouse.id" column of your input data. Required.}

\item{time.var}{The name of the "day" column of your input data. Required.}

\item{tv.var}{The name of the "tumor volume" column of your input data. Required.}

\item{ci}{a logical variable to do bootstrapping for confidence interval}

\item{type}{calculate eGR "ratio" or "difference"}
}
\value{
a data frame containing eGR ratio/difference for each treatment group
}
\description{
get eGR ratio or difference within a trial
}
