% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Model_TGI.R
\name{get_Model_TGI}
\alias{get_Model_TGI}
\title{get tumor growth inhibition TGI for treatment groups}
\usage{
get_Model_TGI(
  tv.data,
  ref.group,
  def = 3,
  group.id.var,
  mouse.id.var,
  time.var,
  tv.var,
  ci = F
)
}
\arguments{
\item{tv.data}{tumor volume data in long format, containing columns "Group","Mouse","Day","TV"}

\item{ref.group}{the reference group in eGR ratio or difference}

\item{def}{definition of TGI. 1: TV_trt/TV_ctl, 2: DeltaTV_trt/DeltaTV_ctl, 3: 1-DeltaTV_trt/DeltaTV_ctl, 4:RTV_trt/RTV_ctl, 5: 1-TV_trt/TV_ctl, 6: 1-RTV_trt/RTV_ctl}

\item{group.id.var}{The name of the "group" column of your input data. Required.}

\item{mouse.id.var}{The name of the "mouse.id" column of your input data. Required.}

\item{time.var}{The name of the "day" column of your input data. Required.}

\item{tv.var}{The name of the "tumor volume" column of your input data. Required.}

\item{ci}{a logical variable to do bootstrapping for confidence interval}
}
\value{
a tibble containing TGI for each treatment group
}
\description{
get tumor growth inhibition TGI for treatment groups
}
