% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geteGRRatio.R
\name{geteGRRatio}
\alias{geteGRRatio}
\title{get eGR ratio for treatment groups}
\usage{
geteGRRatio(
  df,
  grp,
  ref_grp,
  type = c("difference", "ratio"),
  ci = F,
  nrep = 1000
)
}
\arguments{
\item{df}{a data frame containing calculated eGR}

\item{grp}{a specific treatment group}

\item{ref_grp}{the reference group}

\item{type}{calculate eGR ratio or difference}

\item{ci}{a logical value to bootstrap confidence interval or not}

\item{nrep}{the number of bootstrapping for confidence interval}
}
\value{
a list of eGR information
}
\description{
get eGR ratio for treatment groups
}
