% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmFit.R
\name{lmmFit}
\alias{lmmFit}
\title{Fit by linear mixed-effects model (LMM)}
\usage{
lmmFit(tv.data, ref.group, group.id.var, mouse.id.var)
}
\arguments{
\item{tv.data}{Tumor volume data in "long format", that is, tumor volumes are in separate lines, with a Day column. Required.}

\item{ref.group}{The reference group, usually a control/vehicle group for the drug group to compare with. Required.}

\item{group.id.var}{The name of the "group" column of your input data. Required.}

\item{mouse.id.var}{The name of the "mouse.id" column of your input data. Required.}
}
\value{
An S3 object of class "lmmfit" which stores coefficient tables and an S4 object of class "lmerModLmerTest".
}
\description{
This function uses linear mixed model to test if there are significant differences in growth rates between treatment groups.
Need to set reference group. Results will be
in the simplest case, random slope for each mouse.
}
\details{
Assuming tumor grows exponentially, log-transformed tumor growth curve will become a straight line versus time. LMM
is comparing the slopes of the lines, and will generate a p-value indicating whether there are significant differences in the slopes.
}
