% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gamfit.R
\name{plot.gamfit}
\alias{plot.gamfit}
\title{Plot GAM fitted tumor growth curves}
\usage{
\method{plot}{gamfit}(object, show.points = T)
}
\arguments{
\item{object}{an object returned by gamFit()}

\item{show.points}{a logical value to decide whether to show the original data points or not.}
}
\description{
This function visualized the GAM fitted tumor growth curves, by using the function is plot_smooths() in package tidymv
}
