% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lmmfit.R
\name{plot.lmmfit}
\alias{plot.lmmfit}
\title{Plot LMM fitted tumor growth curves}
\usage{
\method{plot}{lmmfit}(object, show.points = T)
}
\arguments{
\item{object}{an object returned by lmmFit()}

\item{show.points}{a logical value to decide whether to show the original data points or not.}
}
\value{
a ggplot2 object of plot
}
\description{
This function visualized the LMM fitted tumor growth curves, by using the ggpredict() function from ggeffects package.
}
