% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgi_group.R
\name{tgi_group}
\alias{tgi_group}
\title{Calculate tumor growth inhibition for treatment groups}
\usage{
tgi_group(
  tv.data,
  ref.group,
  def = 3,
  group.id.var,
  mouse.id.var,
  time.var,
  tv.var
)
}
\arguments{
\item{tv.data}{a data frame containing tumor volume data in long format}

\item{ref.group}{the reference group}

\item{def}{definition of TGI}

\item{group.id.var}{The name of the "group" column of your input data. Required.}

\item{mouse.id.var}{The name of the "mouse.id" column of your input data. Required.}

\item{time.var}{The name of the "day" column of your input data. Required.}

\item{tv.var}{The name of the "tumor volume" column of your input data. Required.}
}
\value{
a list of TGI information
}
\description{
Calculate tumor growth inhibition for treatment groups
}
\details{
Definitions of TGI:
1. TV_t/TV_c
2. DeltaTV_t/DeltaTV_c
3. 1 - DeltaTV_t/DeltaTV_c
4. RTV_t/RTV_c
5. 1 - TV_t/TV_c
6. 1 - RTV_t/RTV_c
}
