% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transposes gene expression data from thin (each row is a gene) to fat (each row is a sample)}
\usage{
transpose(data.thin, gene.id)
}
\arguments{
\item{gene.id}{a vector of gene ID strings that must be unique and matches the sequence of the genes.}

\item{data.wide}{gene expression data frame or matrix in the "thin" format, in which each row is a gene.
The data should be all numeric, and exclude the gene id column.}
}
\value{
returns a data frame with the first column as Model and the rest columns as gene expression data
}
\description{
Transposes gene expression data from thin (each row is a gene) to fat (each row is a sample)
}
