% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{Converts tumor growth data from "wide" format to "long" format}
\usage{
wide2long(tv_data, id.vars)
}
\arguments{
\item{tv_data}{Tumor volume data frame in "wide" format, that is, each row is a mouse and each column is a time point. In addition, need to have a Group column, indicating each mouse's group information.}

\item{id.vars}{The "ID" columns that describes mouse information, such as Mouse ID, Group information, and everything except Day columns. Can use a vector of strings, or indices of the columns in the data table.}
}
\value{
A long data formatted data frame
}
\description{
Most tumor growth data are in "wide" format, in which each row represents a mouse, and each column represents a time point.
This format cannot be read by lmmFit and other model fitting functions, and has to be converted to long format.
}
