% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgi_group.R
\name{tgi_group}
\alias{tgi_group}
\title{Calculate tumor growth inhibition for treatment groups}
\usage{
tgi_group(
  tv.data,
  ref.group,
  grp,
  type = c("geometric", "arithmetic"),
  def = 6,
  ci = F,
  bound = 90
)
}
\arguments{
\item{tv.data}{a data frame containing tumor volume data in long format}

\item{ref.group}{the reference group}

\item{grp}{the group to calculate TGI for}

\item{type}{the type of the mean of TGI, choose from "arithmetic" and "geometric".}

\item{def}{definition of TGI}

\item{ci}{whether to do 1000 times bootstrapping to determine the 95 confidence interval for the calculated TGI.}

\item{bound}{a number for the confidence interval boundary. Suggest 90 or 95.}
}
\value{
a list of TGI information
}
\description{
Calculate tumor growth inhibition for treatment groups
}
\details{
Definitions of TGI:
1. TV_t/TV_c
2. DeltaTV_t/DeltaTV_c
3. 1 - DeltaTV_t/DeltaTV_c
4. RTV_t/RTV_c
5. 1 - TV_t/TV_c
6. 1 - RTV_t/RTV_c
}
